requirements.txt
fastapi
uvicorn
pydantic
langchain-openai
browser-use
python-dotenv
aiofiles
httpx
